+++ src/css/style.css
@@ -210,7 +210,7 @@ button {
 #cart-count {
   position: absolute;
   top: 10px;
-  right: 0;
+  right: -10px;
   display: inline-flex;
   justify-content: center;
   align-items: center;
