+++ src/index.html
@@ -36,6 +36,5 @@
       </section>
     </main>
     <footer id="main-footer"></footer>
-    <script src="js/main.js" type="module"></script>
   </body>
 </html>
