+++ src/js/cartCount.mjs
@@ -1,22 +1,23 @@
-import { getLocalStorage } from "./utils.mjs";
+//USED TO WORK 
 
-const span = document.createElement("span");
-span.id = "cart-count";
-const header = document.querySelector("#main-header");
-header.appendChild(span);
+// import { getLocalStorage } from "./utils.mjs";
 
 
-
-export async function displayCounter() {
-  let cartItems = await getLocalStorage("so-cart");
-  if (!cartItems) {
-    const element = document.querySelector("#cart-count");
-    element.textContent = 0;
-  } else {
-    const element = document.querySelector("#cart-count");
-    cartItems = getLocalStorage("so-cart").length;
-    element.textContent = cartItems;
+// export function displayCounter() {
+//     window.onload = () => {
+//         const span = document.createElement("span");
+//         span.id = "cart-count";
+//         const div = document.querySelector(".cart");
+//         div.appendChild(span);
     
-  }
-
-}
+//     let cartItems = getLocalStorage("so-cart");
+//     if (!cartItems) {
+//           const element = document.querySelector("#cart-count");
+//           element.textContent = 0;
+//         } else {
+//               const element = document.querySelector("#cart-count");
+//               cartItems = getLocalStorage("so-cart").length;
+//               element.textContent = cartItems;
+//         }            
+//     }
+// }
