+++ src/js/product.js
@@ -2,7 +2,6 @@ import { getParam } from "./utils.mjs";
 import ProductData from "./ProductData.mjs";
 import ProductDetails from "./ProductDetails.mjs";
 import { loadHeaderFooter } from "./utils.mjs";
-import { displayCounter } from "./cartCount.mjs";
 
 loadHeaderFooter();
 
@@ -13,5 +12,3 @@ const productId = getParam("product");
 
 const product = new ProductDetails(productId, dataSource);
 product.init();
-
-displayCounter();
