+++ src/public/partials/header.html
@@ -28,4 +28,5 @@
           <text x="0" y="120" fill="#000000" font-size="5px" font-weight="bold" font-family="'Helvetica Neue', Helvetica, Arial-Unicode, Arial, Sans-serif">from the Noun Project</text> -->
     </svg>
   </a>
+  <span id="cart-count"></span>
 </div>
